\name{delete.duplicates}
\alias{delete.duplicates}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Delete duplicated state events}
\description{
Used internally by cowAnalyze
}
\usage{
delete.duplicates(obs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obs}{A dataframe containing state events}
}
\details{

}

\author{Matti Pastell}
