\name{freq.count}
\alias{freq.count}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Count frequency of behaviors }
\description{
Used internally by cowAnalyze
}
\usage{
freq.count(x)
}

\arguments{
  \item{x}{A numeric or factor vector of behaviors}
  }


\author{Matti Pastell}